+++ Assets/Prefabs/Pickup.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 9058588071734529034}
   - component: {fileID: 9058588071734529525}
   - component: {fileID: 2255440775196092560}
-  - component: {fileID: 4916404529471799593}
   m_Layer: 0
   m_Name: Pickup
   m_TagString: Pickup
@@ -114,17 +113,3 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &4916404529471799593
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9058588071734529527}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 10ef676dba5e8a94fa10ae04e57fc9aa, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  pickupObject: {fileID: 9058588071734529527}
-  Player: {fileID: 6409904791983566118, guid: d84c1522a6661364f807d895501a8878, type: 3}
