--- Assets/Scripts/Ammo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c33071bdec570e147910dcb2cbacd8c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
