--- Assets/Scripts/ClampHealthBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f02130c4f00a8e148b5932147d35d25d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
