+++ Assets/Scripts/HealthScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4aaa0272a232e44f874c2a0201b0563
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
