+++ Assets/Scripts/PlayerController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4f35afaf50174245b722e43d60e76f0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
