+++ Assets/Scripts/ShootingScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a950f87413053f94f88395ef0ae558f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
