+++ Assets/Scripts/ShootingScripts/Pickup.cs.meta
