+++ Assets/Scripts/ShootingScripts/Sniper.cs
